requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
colorama==0.4.6
discord.py==2.4.0
exceptiongroup==1.2.2
frozenlist==1.4.1
idna==3.7
iniconfig==2.0.0
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.3.1
python-dotenv==1.0.1
tomli==2.0.1
yarl==1.9.4
